Rename exernal secondary instances
