+++ src/jmh/java/org/javarosa/benchmarks/BenchmarkTemplate.java
@@ -1,8 +1,8 @@
 package org.javarosa.benchmarks;
 
 import static java.nio.file.Files.readAllBytes;
-import static org.javarosa.benchmarks.BenchmarkUtils.dryRun;
-import static org.javarosa.benchmarks.BenchmarkUtils.prepareAssets;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.dryRun;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.prepareAssets;
 import static org.javarosa.core.reference.ReferenceManagerTestUtils.setUpSimpleReferenceManager;
 
 import java.io.IOException;
