+++ src/jmh/java/org/javarosa/benchmarks/FormParserHelperParseExternalInstanceBenchmark.java
@@ -1,7 +1,7 @@
 package org.javarosa.benchmarks;
 
-import static org.javarosa.benchmarks.BenchmarkUtils.dryRun;
-import static org.javarosa.benchmarks.BenchmarkUtils.getNigeriaWardsXMLWithExternal2ndryInstance;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.dryRun;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.getNigeriaWardsXMLWithExternal2ndryInstance;
 
 import java.io.IOException;
 import java.nio.file.Path;
