+++ src/jmh/java/org/javarosa/benchmarks/PopulateTreeNodeBenchmark.java
@@ -1,7 +1,7 @@
 package org.javarosa.benchmarks;
 
-import static org.javarosa.benchmarks.BenchmarkUtils.dryRun;
-import static org.javarosa.benchmarks.BenchmarkUtils.prepareAssets;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.dryRun;
+import static org.javarosa.benchmarks.utils.BenchmarkUtils.prepareAssets;
 
 import java.io.IOException;
 import java.nio.file.Files;
